docs: Modify comments

+++ api/src/main/java/com/thesurvey/api/service/converter/CertificationTypeConverter.java
@@ -8,31 +8,32 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class CertificationTypeConverter {
 
-    /*
-     * Integer is set to the index of EnumTypeEntity.CertificationType.
-     * value : CertificationType
-     * 0: KAKAO, 1: NAVER, 2: GOOGLE, 3: WEBMAIL, 4: DRIVER_LICENSE, 5: MOBILE_PHONE
+    /**
+     * return List<Integer> certificationTypeList to
+     * List<CertificationType> convertedCertificationTypeList.
+     * @param certificationTypeList List converted from CertificationType to Integer in db.
+     * @return convertedCertificationTypeList List converted from Integer to CertificationType.
      */
     public List<CertificationType> toCertificationTypeList(List<Integer> certificationTypeList) {
         List<CertificationType> convertedCertificationTypeList = new ArrayList<>();
         for (Integer type : certificationTypeList) {
 
-            if (type.intValue() == 0) {
+            if (type == 0) {
                 convertedCertificationTypeList.add(CertificationType.KAKAO);
             }
-            else if (type.intValue() == 1) {
+            else if (type == 1) {
                 convertedCertificationTypeList.add(CertificationType.NAVER);
             }
-            else if (type.intValue() == 2) {
+            else if (type == 2) {
                 convertedCertificationTypeList.add(CertificationType.GOOGLE);
             }
-            else if (type.intValue() == 3) {
+            else if (type == 3) {
                 convertedCertificationTypeList.add(CertificationType.WEBMAIL);
             }
-            else if (type.intValue() == 4) {
+            else if (type == 4) {
                 convertedCertificationTypeList.add(CertificationType.DRIVER_LICENSE);
             }
-            else if (type.intValue() == 5) {
+            else if (type == 5) {
                 convertedCertificationTypeList.add(CertificationType.MOBILE_PHONE);
             }
